requirements.txt
fastapi
uvicorn
langgraph
langchain
openai
pydantic
python-dotenv
requests